Equipment tweaks and linting.

+++ src/macros/ability-execution/bodyguard-backup-roll-feat-save.mjs
@@ -1,8 +1,5 @@
 const data = /** @type {Teriock.HookData.RollFeatSave} */ scope.data;
-if ([
-  "mov",
-  "str",
-].includes(data.attribute)) {
+if (["mov", "str"].includes(data.attribute)) {
   data.options.advantage = true;
   const effect = actor.effects.getName("Bodyguard Backup Effect");
   await effect.delete();
